Adds world folder
This folder is actually a part of a normal Evennia game, and was
initially excluded under the false assumption that everything in it
would be story elements excluded from the public code. Instead, all
excluded story-related code for gridpunx will be added to the
.gitignore file as it is created. Note that this does not include the
default build script that is generated by Evennia.