Adds 'receive' locks to some objects
Description:
This adds a lock string of 'receive:true()' to the following
classes: characters.Character, npc.RealTalkingNPC, and the newly
createdRealContainer. These are the only objects that make sense
to allow giving other objects to. Containers may utilize a
different command in the future, but it could still make use
of the same 'receive' lock.

Background:
The modified give command checks for an access_type of 'receive'
when it is called. If an object is not explicitly assigned this
access in a lock string, then you cannot assign it as a target in
a '@give' command.

Notes:
This fixes issue  #11