Fixes potential performance issues with HarshClimate script
This commit changes the HarshClimate script by setting the start_delay
value to True, moving the interval assignement into the at_start hook,
and changing the interval to a random value between 1-2 minutes.

These changes are meant to avoid potential performance issues. If
the game has a lot of rooms with this script attached, then all
of them updating at the same interval could cause problems.