Adds 'AuthenticatedExit' typeclass to exits.py
The AuthenticatedExit typeclass is for creating doors wich are locked
based on properties of an object with a typeclass of:
'physical.things.IdentityToken'

In order for players to traverse the exit, they must run the 'use'
command on the door and specify an IdentityToken type object as their
<inventory object> ('=' or '-with' parameter). Builders can configure
the exit's 'use_mode' attribute to change how the door handles the
identity tokens that are used with it. Only Admins and higher can view
and edit the sensitive 'granted_keys' attribute, which is a list of
hashed key values.