Fixes AttributeError in AuthenticatedExit
This fixes a problem where an AuthenticatedExit object would throw
AttributeError messages if a player issued a 'use' command on the Exit
without specifing the '=' or '-with' parameter at all. The solution
was to check the 'with_obj' before running the authentication checks
and end the process if it was None.