Adds token auth to RealContainer class
This implements token authentication in the RealContainer class.
The option is defined in game by setting the object's 'use_mode'
attribute to a string value of 'token'. Setting this mode utilizes a new
auth_common.py module that provides a common set of functions that
other objects can use to provide authenticated access based on
IdentityToken objects. Token mode requires the object to have an
additional attibute named 'auth_mode' that determines what happens
when it is used with a token. By default, these values are set to
'unlocked' and 'allowing' when an object is created/updated.